Un-nested frontend folder
The folder containing frontend files used to be in "class-compass" inside the "Frontend Stage" folder.
Moved "class-compass" into the main directory, renamed it "frontend" and removed the empty "Frontend Stage" folder.
Now backend and frontend folders are in the same directory with direct access to the files in each folder instead of being nested in another folder.