Created Home Page
Created a home page that fetches all courses and displays each course's information.

Course information is currently fetched from the test data in the Public directory since the backend API is not yet created. Once the backend API is created, this fetch will need to be updated to fetch the data from the backend API.

Courses are currently displayed in plaintext. This will need to be changed to some kind of Course Node component which has yet to be created. I(Ryu) am thinking of using "react flow" for this but haven't gotten around to learning it and seeing its viability yet.