Create Course Button and Popup components
Created Course button component that displays just the title of a course, and on click, opens a course information popup that displays the rest of the course information.

The popup also has a button to "View Prerequisite Tree" which currently doesn't do anything but will eventually link to a page that fetches specifically that course, its prerequisites, and courses that don't list it as prerequisite and displays it as a tree.