Node Tree Changes
The prerequisites tree now works with the new course data format. However, currently when the situation where you can take multiple different courses to fill the same prerequisite requirement occurs, the node tree simply uses the first listed course. This will be eventually fixed. But I chose to do this for simplicity for now. Also, now when you click on a course in a prerequisite tree it will direct you to the prerequisite tree for the course you clicked.