Added handling for non-fetched courses
Added handling for courses not found in the database. Before, if a prerequisite course was not found in the course list fetched from the database, it would return null, resulting in runtime errors. Now it returns a dummy course object with the prerequisite string as the course code and name, so that no runtime errors occur, and the course can still be displayed despite its data not being fetched.