Changed button to display course code as type and number
Change how the course button displays the course code.
Previously it would display it as it is formatted in the database (eg. COMPSCI220)

But now course button calls the splitCourseCode function that splits the code into two tokens; one being the course type (eg. COMPSCI) and the other being course number (eg. 220).

This way the course code text is a bit more dynamic, able to wrap around if it is too long, so that course button can be resized easier with less styling issues.