Updated fetchAllCoursesJSON and added getFilteredCourses function
Updated fetchAllCoursesJSON and added getFilteredCourses function.

fetchAllCoursesJSON now returns the courses it fetched, instead of updated a state using a state updating method passed into it as a parameter. This is to accommodate changes with how the homepage works with the search bar to reduce api calls to improve performance.

getFilteredCourses was pulled from the search bar component to modularize the code and make it easier to test. Function was also modified slightly to be a bit more readable and output status to console.