Created functions to add and remove courses taken from student accounts
Creates functions that make PUT requests to the backend to update a specific student's courses_taken field.

One function adds a course passed in as a parameter to the student's courses_taken list.

The other function removes a course passed in as a parameter to the student's courses_taken list.

These functions exist for use by the course popup component to allow a logged in student to mark a course as taken or not taken, update the databse accordingly, and to be reflected visually through color coding.