Added functions getCookie and getUser
Added the function getCookie that parses the cookies of the site and returns the value for a specified key.

Added a function that accordingly fetches a student or advisor account by id. Currently a route for getting an advisor by id does not exist so code for that is commented out, and needs to be corrected once the backend route is fixed.