Added new states for color coding based on logged in student
Added new state to store user fetched using functions from getUser and passing in parameters retrieved from the cookies using getCookies.

Currently the backend does not send back user._id in the handleSignIn functions for both student and advisor and the login and register pages do not set those values along with userType as cookies, so fetch functionality would give an error and is thus commented out currently.

Once those changes to the backend and login/register pages are made to allow all the necessary information to be retrieved, the code calling functions to getUser and setUser can be uncommented out, and color coding functionality should be working.