Added color coding
Added color coding for course buttons by passing in user object and checking if course is in user's courses_taken list (if they have one. Depending on if it is taken or not, the id of the button is set accordingly, which the css applies different background color styles to.

Currently if a course is not taken, it is grey.
If the course is taken, it is green.

The course button also passes user information to the CourseInfoPopup component, to allow it to show certain buttons accordingly.