Added buttons to mark courses as taken or not
Added the ability to mark courses as taken or not if signed in as student, which will update the student's courses_taken field and will color code the course list accordingly.

The proper buttons will only appear based on if the student is signed in as a student and if the course is taken or not.
If the user is signed in as a student and they have taken the course, they only see the button to set course as not taken.
If the user is signed in as a student and they have not taken the course, they only see the button to set the course as taken.

Also added a button for advisors to edit that course taking them to a page where they will eventually be able to fill out a form to edit the course and update it on the database.

This button will only appear if the user is logged in as an advisor.