Removed option for advisor login and register
Removed option for advisor login and register by commenting out advisor option of drop down menu. This is because for 1.0 we are pushing advisor functionality to reach, and as there is no reason to register and login as an advisor for 1.0, there wouldn't be any reason to allow that login/register option.